# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Aqar Data Studio — Professional Real Estate Data Management Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /lookup-options:
    get:
      operationId: listLookupOptions
      tags: [lookup-options]
      summary: List lookup options, optionally filtered by category
      parameters:
        - name: category
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Lookup options list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LookupOption"

    post:
      operationId: createLookupOption
      tags: [lookup-options]
      summary: Create a new lookup option
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LookupOptionInput"
      responses:
        "201":
          description: Created lookup option
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LookupOption"
  /lookup-options/{id}:
    patch:
      operationId: updateLookupOption
      tags: [lookup-options]
      summary: Update a lookup option
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LookupOptionUpdate"
      responses:
        "200":
          description: Updated lookup option
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LookupOption"

    delete:
      operationId: deleteLookupOption
      tags: [lookup-options]
      summary: Delete a lookup option
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"

  # ─── DASHBOARD ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    DeleteResult:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
        id: { type: string }

    # ── Property ────────────────────────────────────────────────────────────
...
    LookupOption:
      type: object
      required: [id, category, value, label, active, sortOrder]
      properties:
        id: { type: string }
        category: { type: string }
        value: { type: string }
        label: { type: string }
        active: { type: boolean }
        sortOrder: { type: number }
    LookupOptionInput:
      type: object
      required: [category, value, label]
      properties:
        category: { type: string }
        value: { type: string }
        label: { type: string }
        active: { type: boolean }
        sortOrder: { type: number }
    LookupOptionUpdate:
      type: object
      properties:
        value: { type: string }
        label: { type: string }
        active: { type: boolean }
        sortOrder: { type: number }

    # ── Dashboard ─────────────────────────────────────────────────────────
...